Fix PR gates for media overlay flow
