+++ packages/studio-ui-react/src/media-intake-panel.tsx
@@ -22,6 +22,142 @@ export type MediaIntakePanelProps = Readonly<{
 
 const defaultIsSupportedUploadFile = () => true;
 
+const useMediaIntakeInteraction = ({
+  isBusy,
+  isSupportedUploadFile,
+  onFileSelected,
+}: Readonly<{
+  isBusy: boolean;
+  isSupportedUploadFile: (file: File) => boolean;
+  onFileSelected: (file: File) => void;
+}>) => {
+  const [isDragActive, setIsDragActive] = React.useState(false);
+
+  const handleDragEnter = React.useCallback(
+    (event: React.DragEvent<HTMLElement>) => {
+      event.preventDefault();
+      if (!isBusy) {
+        setIsDragActive(true);
+      }
+    },
+    [isBusy]
+  );
+
+  const handleDragOver = React.useCallback(
+    (event: React.DragEvent<HTMLElement>) => {
+      event.preventDefault();
+      if (!isBusy) {
+        event.dataTransfer.dropEffect = 'copy';
+      }
+    },
+    [isBusy]
+  );
+
+  const handleDragLeave = React.useCallback((event: React.DragEvent<HTMLElement>) => {
+    event.preventDefault();
+    if (!event.currentTarget.contains(event.relatedTarget as Node | null)) {
+      setIsDragActive(false);
+    }
+  }, []);
+
+  const handleDrop = React.useCallback(
+    (event: React.DragEvent<HTMLElement>) => {
+      event.preventDefault();
+      setIsDragActive(false);
+      if (isBusy) {
+        return;
+      }
+      const file = event.dataTransfer.files?.[0];
+      if (file && isSupportedUploadFile(file)) {
+        onFileSelected(file);
+      }
+    },
+    [isBusy, isSupportedUploadFile, onFileSelected]
+  );
+
+  return {
+    isDragActive,
+    handleDragEnter,
+    handleDragOver,
+    handleDragLeave,
+    handleDrop,
+  } as const;
+};
+
+const MediaIntakePanelBody = ({
+  accept,
+  browseActionLabel,
+  description,
+  inputTestId,
+  inputRef,
+  isBusy,
+  isSupportedUploadFile,
+  onFileSelected,
+  statusClassName,
+  statusMessage,
+  supportLabel,
+  title,
+}: Readonly<{
+  accept?: string;
+  browseActionLabel: string;
+  description: string;
+  inputTestId?: string;
+  inputRef: React.RefObject<HTMLInputElement | null>;
+  isBusy: boolean;
+  isSupportedUploadFile: (file: File) => boolean;
+  onFileSelected: (file: File) => void;
+  statusClassName: string;
+  statusMessage?: string | null;
+  supportLabel: string;
+  title: string;
+}>) => (
+  <div className="mx-auto flex max-w-3xl flex-col items-center gap-4 text-center">
+    <div className="flex h-16 w-16 items-center justify-center rounded-full border border-border/70 bg-muted/40 text-muted-foreground">
+      <Upload aria-hidden="true" className="h-8 w-8" strokeWidth={2.2} />
+    </div>
+
+    <div className="space-y-2">
+      <h2 className="text-2xl font-semibold tracking-tight text-foreground">{title}</h2>
+      <p className="max-w-2xl text-sm leading-6 text-muted-foreground">{description}</p>
+    </div>
+
+    <input
+      ref={inputRef}
+      hidden
+      accept={accept}
+      data-testid={inputTestId}
+      type="file"
+      onChange={(event) => {
+        const file = event.currentTarget.files?.[0];
+        if (file && isSupportedUploadFile(file)) {
+          onFileSelected(file);
+        }
+        event.currentTarget.value = '';
+      }}
+    />
+
+    <div className="flex flex-col items-center gap-3 sm:flex-row sm:flex-wrap sm:justify-center">
+      <Button
+        className="h-11 rounded-xl px-5 text-sm font-semibold"
+        disabled={isBusy}
+        type="button"
+        onClick={() => inputRef.current?.click()}
+      >
+        {browseActionLabel}
+      </Button>
+
+      <div className="flex flex-wrap items-center justify-center gap-3 text-sm text-muted-foreground">
+        <span>{supportLabel}</span>
+        <div className="flex items-center gap-2 text-foreground/70">
+          <FileImage aria-hidden="true" className="h-4 w-4" />
+        </div>
+      </div>
+    </div>
+
+    {statusMessage ? <p className={`text-sm font-medium ${statusClassName}`}>{statusMessage}</p> : null}
+  </div>
+);
+
 export const MediaIntakePanel = ({
   accept,
   browseActionLabel,
@@ -37,8 +173,12 @@ export const MediaIntakePanel = ({
   title,
 }: MediaIntakePanelProps) => {
   const inputRef = React.useRef<HTMLInputElement | null>(null);
-  const [isDragActive, setIsDragActive] = React.useState(false);
   const isBusy = phase === 'initializing' || phase === 'uploading' || phase === 'finalizing';
+  const { handleDragEnter, handleDragLeave, handleDragOver, handleDrop, isDragActive } = useMediaIntakeInteraction({
+    isBusy,
+    isSupportedUploadFile,
+    onFileSelected,
+  });
 
   const statusClassName =
     statusTone === 'error'
@@ -54,84 +194,25 @@ export const MediaIntakePanel = ({
         isDragActive ? 'border-primary/50 bg-muted/30' : 'border-border/70'
       }`}
       data-testid="media-intake-shelf"
-      onDragEnter={(event) => {
-        event.preventDefault();
-        if (isBusy) {
-          return;
-        }
-        setIsDragActive(true);
-      }}
-      onDragOver={(event) => {
-        event.preventDefault();
-        if (isBusy) {
-          return;
-        }
-        event.dataTransfer.dropEffect = 'copy';
-      }}
-      onDragLeave={(event) => {
-        event.preventDefault();
-        if (event.currentTarget.contains(event.relatedTarget as Node | null)) {
-          return;
-        }
-        setIsDragActive(false);
-      }}
-      onDrop={(event) => {
-        event.preventDefault();
-        setIsDragActive(false);
-        if (isBusy) {
-          return;
-        }
-        const file = event.dataTransfer.files?.[0];
-        if (file && isSupportedUploadFile(file)) {
-          onFileSelected(file);
-        }
-      }}
+      onDragEnter={handleDragEnter}
+      onDragOver={handleDragOver}
+      onDragLeave={handleDragLeave}
+      onDrop={handleDrop}
     >
-      <div className="mx-auto flex max-w-3xl flex-col items-center gap-4 text-center">
-        <div className="flex h-16 w-16 items-center justify-center rounded-full border border-border/70 bg-muted/40 text-muted-foreground">
-          <Upload aria-hidden="true" className="h-8 w-8" strokeWidth={2.2} />
-        </div>
-
-        <div className="space-y-2">
-          <h2 className="text-2xl font-semibold tracking-tight text-foreground">{title}</h2>
-          <p className="max-w-2xl text-sm leading-6 text-muted-foreground">{description}</p>
-        </div>
-
-        <input
-          ref={inputRef}
-          hidden
-          accept={accept}
-          data-testid={inputTestId}
-          type="file"
-          onChange={(event) => {
-            const file = event.currentTarget.files?.[0];
-            if (file && isSupportedUploadFile(file)) {
-              onFileSelected(file);
-            }
-            event.currentTarget.value = '';
-          }}
-        />
-
-        <div className="flex flex-col items-center gap-3 sm:flex-row sm:flex-wrap sm:justify-center">
-          <Button
-            className="h-11 rounded-xl px-5 text-sm font-semibold"
-            disabled={isBusy}
-            type="button"
-            onClick={() => inputRef.current?.click()}
-          >
-            {browseActionLabel}
-          </Button>
-
-          <div className="flex flex-wrap items-center justify-center gap-3 text-sm text-muted-foreground">
-            <span>{supportLabel}</span>
-            <div className="flex items-center gap-2 text-foreground/70">
-              <FileImage aria-hidden="true" className="h-4 w-4" />
-            </div>
-          </div>
-        </div>
-
-        {statusMessage ? <p className={`text-sm font-medium ${statusClassName}`}>{statusMessage}</p> : null}
-      </div>
+      <MediaIntakePanelBody
+        accept={accept}
+        browseActionLabel={browseActionLabel}
+        description={description}
+        inputRef={inputRef}
+        inputTestId={inputTestId}
+        isBusy={isBusy}
+        isSupportedUploadFile={isSupportedUploadFile}
+        onFileSelected={onFileSelected}
+        statusClassName={statusClassName}
+        statusMessage={statusMessage}
+        supportLabel={supportLabel}
+        title={title}
+      />
     </section>
   );
 };
